Move the tui.App.updateKeyTextView func to core/tui/keys.go
Unclutter the core/tui/main.go file.
